name change; and updated to newer version provided by Colin
